Enhance entity command handler with client connection
Pass the WS connection as an optional command parameter to the client.
This allows the client to send directed WS events back. Otherwise, the
only option is to use event broadcasts to all connected clients.
Implementation is backward compatible to existing clients without the
new websocket parameter in the CommandHandler callback, or extending a
concrete Entity class.

Import cleanup to avoid circular imports through the ucapi definition.